Convert to Jotai atoms for state
